requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
line-bot-sdk==2.4.2
openai==0.28.1
MarkupSafe>=2.1.1
gunicorn==20.1.0
tenacity==5.1.5 
mega.py==1.0.8
langchain==0.0.69
faiss-cpu
sentence-transformers==2.2.2
huggingface_hub==0.14.1







